Named-based schema matching between table and partitions (#37)
This change allows the possibility of using names to match schemas between a partition and a table and patches parquet implementation to make it compatible with it. This behavior is disabled by default and needs the usage of a configuration variable. This has only be tested with parquet datasets.

The set of changes is the following:

Add a new behavior to the StructConverter to allow matching by name instead of by index. The activation of this behavior is done through a configuration variable, so the ObjectInspectorConverters API has been changed to allow the passing of the config.
Modify the ParquetHiveSerDe to shape the data coming from the reader with the schema of the partition, as this is expected by the converter in the operator.