Support column statistics when partition values contain leading 0s
When analyzing columns, Hive computes stats correctly but fails to
insert them in the Metastore when the partition contains values
with leading zeros (eg. day=2020-05-04/hour=01 where hour is
declared as integer).
This change uses constant literals as the partition values when
inserting statistics, allowing to have formatted numerals as
partition values.